+++ src/js/constants.js
@@ -1,5 +1,7 @@
-export const IN_BROWSER = typeof window !== 'undefined';
-export const WINDOW = IN_BROWSER ? window : {};
+export const IS_BROWSER = typeof window !== 'undefined';
+export const WINDOW = IS_BROWSER ? window : {};
+export const IS_TOUCH_DEVICE = IS_BROWSER ? 'ontouchstart' in WINDOW.document.documentElement : false;
+export const HAS_POINTER_EVENT = IS_BROWSER ? 'PointerEvent' in WINDOW : false;
 export const NAMESPACE = 'viewer';
 
 // Actions
@@ -34,9 +36,12 @@ export const EVENT_HIDDEN = 'hidden';
 export const EVENT_HIDE = 'hide';
 export const EVENT_KEY_DOWN = 'keydown';
 export const EVENT_LOAD = 'load';
-export const EVENT_POINTER_DOWN = WINDOW.PointerEvent ? 'pointerdown' : 'touchstart mousedown';
-export const EVENT_POINTER_MOVE = WINDOW.PointerEvent ? 'pointermove' : 'touchmove mousemove';
-export const EVENT_POINTER_UP = WINDOW.PointerEvent ? 'pointerup pointercancel' : 'touchend touchcancel mouseup';
+export const EVENT_TOUCH_START = IS_TOUCH_DEVICE ? 'touchstart' : 'mousedown';
+export const EVENT_TOUCH_MOVE = IS_TOUCH_DEVICE ? 'touchmove' : 'mousemove';
+export const EVENT_TOUCH_END = IS_TOUCH_DEVICE ? 'touchend touchcancel' : 'mouseup';
+export const EVENT_POINTER_DOWN = HAS_POINTER_EVENT ? 'pointerdown' : EVENT_TOUCH_START;
+export const EVENT_POINTER_MOVE = HAS_POINTER_EVENT ? 'pointermove' : EVENT_TOUCH_MOVE;
+export const EVENT_POINTER_UP = HAS_POINTER_EVENT ? 'pointerup pointercancel' : EVENT_TOUCH_END;
 export const EVENT_READY = 'ready';
 export const EVENT_RESIZE = 'resize';
 export const EVENT_SHOW = 'show';
@@ -50,6 +55,11 @@ export const EVENT_ZOOMED = 'zoomed';
 
 // Data keys
 export const DATA_ACTION = `${NAMESPACE}Action`;
+
+// RegExps
+export const REGEXP_SPACES = /\s\s*/;
+
+// Misc
 export const BUTTONS = [
   'zoom-in',
   'zoom-out',
@@ -63,6 +73,3 @@ export const BUTTONS = [
   'flip-horizontal',
   'flip-vertical',
 ];
-
-// RegExps
-export const REGEXP_SPACES = /\s\s*/;