+++ src/js/utilities.js
@@ -2,7 +2,7 @@ import {
   CLASS_HIDE_MD_DOWN,
   CLASS_HIDE_SM_DOWN,
   CLASS_HIDE_XS_DOWN,
-  IN_BROWSER,
+  IS_BROWSER,
   REGEXP_SPACES,
   WINDOW,
 } from './constants';
@@ -65,7 +65,7 @@ export function isPlainObject(value) {
     const { prototype } = constructor;
 
     return constructor && prototype && hasOwnProperty.call(prototype, 'isPrototypeOf');
-  } catch (e) {
+  } catch (error) {
     return false;
   }
 }
@@ -295,14 +295,14 @@ export function removeData(element, name) {
   if (isObject(element[name])) {
     try {
       delete element[name];
-    } catch (e) {
+    } catch (error) {
       element[name] = undefined;
     }
   } else if (element.dataset) {
     // #128 Safari not allows to delete dataset property
     try {
       delete element.dataset[name];
-    } catch (e) {
+    } catch (error) {
       element.dataset[name] = undefined;
     }
   } else {
@@ -313,7 +313,7 @@ export function removeData(element, name) {
 const onceSupported = (() => {
   let supported = false;
 
-  if (IN_BROWSER) {
+  if (IS_BROWSER) {
     let once = false;
     const listener = () => {};
     const options = Object.defineProperty({}, 'once', {
@@ -615,6 +615,7 @@ export function getPointer({ pageX, pageY }, endOnly) {
   };
 
   return endOnly ? end : assign({
+    timeStamp: Date.now(),
     startX: pageX,
     startY: pageY,
   }, end);